feat: Markup group meeting page
