+++ src/components/templates/index.ts
@@ -1,1 +1,2 @@
 export { default as GoBackHome } from './GoBackHome';
+export { default as GroupPage } from './GroupPage';
